requirements.txt
# Core Dependencies
# New unified SDK (replaces deprecated google-generativeai)
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# GUI Framework
streamlit>=1.39.0

# Utilities
pathlib>=1.0.1
typing-extensions>=4.0.0

# Logging and Monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: For advanced JSON handling
jsonschema>=4.19.0
